# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prosthuti API",
    "version": "1.0.0",
    "description": "Local Laravel API. Cookie sessions; mutations require CSRF. See API.md for payloads and workflow constraints. Response schemas are intentionally not exhaustive."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "prosthuti_session"
      }
    }
  },
  "paths": {
...
    "/attempts": {
      "get": {
        "operationId": "AttemptController_index_get",
        "tags": [
          "attempts"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful JSON response; see integration guide and controller contract."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or role not allowed"
          },
          "404": {
            "description": "Resource unavailable or not owned"
          },
          "422": {
            "description": "Validation error; JSON message and errors"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "parameters": []
      },
      "post": {
        "operationId": "AttemptController_store_post",
        "tags": [
          "attempts"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful JSON response; see integration guide and controller contract."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or role not allowed"
          },
          "404": {
            "description": "Resource unavailable or not owned"
          },
          "422": {
            "description": "Validation error; JSON message and errors"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "419": {
            "description": "Missing or expired CSRF token/session"
          },
          "409": {
            "description": "State or foreign-key conflict"
          },
          "201": {
            "description": "New or idempotently replayed attempt"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/attempts/{attempt}": {
      "get": {
        "operationId": "AttemptController_show_get",
        "tags": [
          "attempts"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful JSON response; see integration guide and controller contract."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or role not allowed"
          },
          "404": {
            "description": "Resource unavailable or not owned"
          },
          "422": {
            "description": "Validation error; JSON message and errors"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "parameters": [
          {
            "name": "attempt",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/attempts/{attempt}/answers": {
      "put": {
        "operationId": "AttemptController_answers_put",
        "tags": [
          "attempts"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful JSON response; see integration guide and controller contract."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or role not allowed"
          },
          "404": {
            "description": "Resource unavailable or not owned"
          },
          "422": {
            "description": "Validation error; JSON message and errors"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "419": {
            "description": "Missing or expired CSRF token/session"
          },
          "409": {
            "description": "State or foreign-key conflict"
          }
        },
        "parameters": [
          {
            "name": "attempt",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/attempts/{attempt}/submit": {
      "post": {
        "operationId": "AttemptController_submit_post",
        "tags": [
          "attempts"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful JSON response; see integration guide and controller contract."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or role not allowed"
          },
          "404": {
            "description": "Resource unavailable or not owned"
          },
          "422": {
            "description": "Validation error; JSON message and errors"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "419": {
            "description": "Missing or expired CSRF token/session"
          },
          "409": {
            "description": "State or foreign-key conflict"
          }
        },
        "parameters": [
          {
            "name": "attempt",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  }
}